+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to Cranberry\CLI will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- Switched formatted Shell::exec output to `Format\String`
+
 ## [0.1.1] - 2016-12-28
 ### Fixed
 - Output returned from Command execution is printed